Implement seeking on streams
Since seeking is a requirement for reading and writing some formats, introduce a disk-backed caching reader and writer that use a temporary file as a swap file of sorts for data as its being streamed. Reads, writes, and seeks are done on the tempfile, and is transferred to/from the underlying stream on demand.

While there will be a performance hit in some cases, there is otherwise no discernable difference now between operations on a file vs on a stream. And since the buffer is backed by a file, memory usage remains low even on large streams.